Changed type to int

+++ 0x1D-binary_trees/14-binary_tree_balance.c
@@ -8,11 +8,12 @@
  */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	size_t height_l;
-	size_t height_r;
+	int height_l;
+	int height_r;
 
 	if (!tree)
 		return (0);
+
 	height_l = tree->left ? 1 + binary_tree_balance(tree->left) : 0;
 	height_r = tree->right ? 1 + binary_tree_balance(tree->right) : 0;
 	return (height_l - height_r);
